# contracts/openapi/commerce.json
{
  "components": {
    "schemas": {
      "AddItemRequest": {
        "properties": {
          "quantity": {
            "default": 1,
            "minimum": 1.0,
            "title": "Quantity",
            "type": "integer"
          },
          "sku_id": {
            "maxLength": 40,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Sku Id",
            "type": "string"
          }
        },
        "required": [
          "sku_id"
        ],
        "title": "AddItemRequest",
        "type": "object"
      },
...
      "CreateCartRequest": {
        "properties": {
          "conversation_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Conversation Id",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          }
        },
        "required": [
          "conversation_id"
        ],
        "title": "CreateCartRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LockRequest": {
        "properties": {
          "order_id": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Order Id",
            "type": "string"
          }
        },
        "required": [
          "order_id"
        ],
        "title": "LockRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "commerce-svc",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/internal/orders/{order_id}/commit": {
      "post": {
        "description": "Payment confirmed: stock leaves inventory, promotion use is final, cart emptied (idempotent).",
        "operationId": "commit_order_internal_orders__order_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Commit Order Internal Orders  Order Id  Commit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit Order",
        "tags": [
          "settlement"
        ]
      }
    },
    "/internal/orders/{order_id}/release": {
      "post": {
        "description": "Payment failed/expired: give stock and promotion uses back (idempotent).",
        "operationId": "release_order_internal_orders__order_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Order Internal Orders  Order Id  Release Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Order",
        "tags": [
          "settlement"
        ]
      }
    },
    "/internal/orders/{order_id}/void": {
      "post": {
        "description": "Paid order refunded (compensation): give promotion uses back; stock is not restocked (idempotent).",
        "operationId": "void_order_internal_orders__order_id__void_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Void Order Internal Orders  Order Id  Void Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Void Order",
        "tags": [
          "settlement"
        ]
      }
    },
    "/internal/quotes/{quote_id}/lock": {
      "post": {
        "description": "Confirm-time: lock the quote to one order and reserve its stock and promotion uses (idempotent).",
        "operationId": "lock_quote_internal_quotes__quote_id__lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "quote_id",
            "required": true,
            "schema": {
              "title": "Quote Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Quote",
        "tags": [
          "settlement"
        ]
      }
    },
    "/v1/carts": {
      "post": {
        "description": "Naturally idempotent: one cart per conversation.",
        "operationId": "get_or_create_cart_v1_carts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Or Create Cart",
        "tags": [
          "carts"
        ]
      }
    },
    "/v1/carts/{cart_id}": {
      "get": {
        "operationId": "get_cart_v1_carts__cart_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cart",
        "tags": [
          "carts"
        ]
      }
    },
    "/v1/carts/{cart_id}/items": {
      "post": {
        "description": "NOT naturally idempotent (it increments), so an Idempotency-Key is required.",
        "operationId": "add_item_v1_carts__cart_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Item",
        "tags": [
          "carts"
        ]
      }
    },
    "/v1/carts/{cart_id}/items/{item_id}": {
      "delete": {
        "operationId": "remove_item_v1_carts__cart_id__items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Item",
        "tags": [
          "carts"
        ]
      },
      "patch": {
        "operationId": "update_item_v1_carts__cart_id__items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Item",
        "tags": [
          "carts"
        ]
      }
    },
    "/v1/carts/{cart_id}/promotion": {
      "delete": {
        "operationId": "remove_promotion_v1_carts__cart_id__promotion_delete",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Promotion",
        "tags": [
          "carts"
        ]
      },
      "put": {
        "operationId": "apply_promotion_v1_carts__cart_id__promotion_put",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyPromoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartSnapshotDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Promotion",
        "tags": [
          "carts"
        ]
      }
    },
    "/v1/carts/{cart_id}/quotes": {
      "post": {
        "operationId": "create_quote_v1_carts__cart_id__quotes_post",
        "parameters": [
          {
            "in": "path",
            "name": "cart_id",
            "required": true,
            "schema": {
              "title": "Cart Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Quote",
        "tags": [
          "carts"
        ]
      }
    },
...
    "/v1/products/search": {
      "get": {
        "operationId": "search_products_v1_products_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "pattern": "^[^\\x00-\\x1f\\x7f]*$",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 24,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_price_minor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price Minor"
            }
          },
          {
            "in": "query",
            "name": "in_stock_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "In Stock Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Products",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/products/{product_id}": {
      "get": {
        "operationId": "get_product_v1_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "currency",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z]{3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSummaryDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}